+++ CHANGELOG.md
@@ -7,8 +7,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- **CUMULUS-2462**
+  - Reingest a granule from granule details page is updated to select a workflow execution arn for reingest.
+  - Batch Reingest and Bulk Reingest Actions are updated to select the workflowName for reingest.
+
+## [v7.0.0]
 
 ## Breaking Changes
+
+This version of the dashboard requires Cumulus API `v9.5.0`
 
 - **CUMULUS-2502**
   - Configuration changes are required to continue connecting the dashboard with Earthdata Metrics
@@ -29,6 +38,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **CUMULUS-2459**
   - Use of ```executions/search-by-granules``` to retrieve executions for a granule/collection
     combination requires minimum CUMULUS API v9.5.0
+
+## All Changes
 
 ### Changed
 
@@ -988,8 +999,8 @@ Fix for serving the dashboard through the Cumulus API.
 ### Added
 
 - Versioning and changelog [CUMULUS-197] by @kkelly51
-
-[Unreleased]: https://github.com/nasa/cumulus-dashboard/compare/v6.0.0...HEAD
+[Unreleased]: https://github.com/nasa/cumulus-dashboard/compare/v7.0.0...HEAD
+[v7.0.0]: https://github.com/nasa/cumulus-dashboard/compare/v6.0.0...v7.0.0
 [v6.0.0]: https://github.com/nasa/cumulus-dashboard/compare/v5.0.0...v6.0.0
 [v5.0.0]: https://github.com/nasa/cumulus-dashboard/compare/v4.0.0...v5.0.0
 [v4.0.0]: https://github.com/nasa/cumulus-dashboard/compare/v3.0.0...v4.0.0